+++ tests/browser/test_export_plan_target_markets.py
@@ -26,6 +26,7 @@ pytestmark = [
 ]
 
 
+@mock.patch.object(exportplan_helpers, 'get_exportplan')
 @mock.patch.object(exportplan_helpers, 'get_comtrade_lastyearimportdata')
 @mock.patch.object(exportplan_helpers, 'get_exportplan_rules_regulations')
 @mock.patch.object(exportplan_helpers, 'get_exportplan_marketdata')
@@ -41,6 +42,7 @@ def test_can_see_target_markets_data(
     mock_get_export_plan_market_data,
     mock_get_export_plan_rules_regulations,
     mock_get_comtrade_last_year_import_data,
+    mock_get_exportplan,
     server_user_browser_dashboard,
     mock_export_plan_dashboard_page_tours,
 ):
@@ -61,6 +63,11 @@ def test_can_see_target_markets_data(
             'Year': 2019,
             'value': 10000,
         }
+    }
+
+    mock_get_exportplan.return_value = {
+        'country': 'Australia', 'commodity_code': '220.850',
+        'target_markets': [{'country': 'China'}], 'rules_regulations': {'country_code': 'CHN'},
     }
 
     live_server, user, browser = server_user_browser_dashboard